forgot one if branch

+++ scripts/data.py
@@ -61,7 +61,7 @@ def extract(num_features, phoneme_feat=False):
         diff = Y.shape[0] - y_phoneme.shape[0]
         if diff != 0:
             if diff % 2 == 0:
-                y_phoneme = np.pad(y_phoneme, (diff/2.0,),  'constant', constant_values=(0,))
+                y_phoneme = np.pad(y_phoneme, (int(diff/2),),  'constant', constant_values=(0,))
             else:
                 y_phoneme = np.pad(y_phoneme, (int(math.floor(diff/2)), int(math.floor(diff/2) + 1)),  'constant', constant_values=(0, 0))
